cs50_finalproject_copilotstructsuggestion/backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
python-dotenv
pytest
pytest-flask